Work on audio stuff

+++ Amiga/Computer/Paula/AudioUnit.cpp
@@ -77,7 +77,7 @@ AudioUnit::getSampleRate()
 }
 
 void
-AudioUnit::setSampleRate(uint32_t rate)
+AudioUnit::setSampleRate(double rate)
 {
 }
